@@ -40,35 +55,52 @@
   },
   "homepage": "https://github.com/CodeTanzania/majifix-common",
   "devDependencies": {
+    "@babel/cli": "^7.5.5",
+    "@babel/core": "^7.5.5",
+    "@babel/node": "^7.5.5",
+    "@babel/preset-env": "^7.5.5",
+    "@babel/register": "^7.5.5",
     "@commitlint/cli": "^8.1.0",
     "@commitlint/config-conventional": "^8.1.0",
     "@commitlint/travis-cli": "^8.1.0",
     "chai": "^4.2.0",
     "commitizen": "^4.0.3",
+    "coveralls": "^3.0.6",
     "cz-conventional-changelog": "^3.0.2",
+    "doxdox": "^3.0.0",
+    "eslint": "^6.1.0",
+    "eslint-config-airbnb-base": "^13.2.0",
+    "eslint-config-prettier": "^6.0.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-jsdoc": "^15.8.0",
+    "eslint-plugin-mocha": "^6.0.0",
+    "eslint-plugin-prettier": "^3.1.0",
     "generate-changelog": "^1.8.0",
-    "grunt": "^1.0.4",
-    "grunt-contrib-jshint": "^2.1.0",
-    "grunt-contrib-watch": "^1.1.0",
-    "grunt-mocha-test": "^0.13.3",
+    "handlebars": "^4.1.2",
     "husky": "^3.0.3",
-    "jshint-stylish": "^2.2.1",
     "lint-staged": "^9.2.1",
     "lodash.template": "^4.5.0",
     "mocha": "^6.2.0",
-    "mongoose": ">=5.6.9",
+    "mongoose": "^5.6.9",
+    "npm-check-updates": "^3.1.20",
+    "nyc": "^14.1.1",
     "prettier": "^1.18.2",
-    "sinon": "^7.4.1",
-    "sinon-chai": "^3.3.0"
+    "rimraf": "^2.6.3",
+    "rollup": "^1.19.4",
+    "shelljs": "^0.8.3"
   },
-  "peerDependencies": {
-    "mongoose": ">=5.6.9"
-  },
+  "peerDependencies": {},
   "dependencies": {
+    "@lykmapipo/common": ">=0.25.0",
     "@lykmapipo/env": ">=0.12.0",
+    "@lykmapipo/mongoose-common": ">=0.25.9",
     "async": ">=3.1.0",
     "lodash": ">=4.17.15",
     "phone": ">=2.3.18"
+  },
+  "engines": {
+    "node": ">=12.4.0",
+    "npm": ">=6.9.0"
   },
   "config": {
     "commitizen": {
